fixed leaks

+++ src/harriet/VariableType.hpp
@@ -14,7 +14,7 @@ namespace harriet {
 class VariableType {
 public:
    /// Each type has a length
-   enum struct Type : uint8_t {TBool, TInteger, TFloat, TCharacter, TUndefined}; // Undefined is needed only for coding. There are some places where we create a value, which has no type yet.
+   enum struct Type : uint16_t {TBool, TInteger, TFloat, TCharacter, TUndefined}; // Undefined is needed only for coding. There are some places where we create a value, which has no type yet.
    Type type;
    uint16_t length;
